+++ elmyra.ip.access.epo/elmyra/ip/access/epo/subscribers.py
@@ -1,6 +1,8 @@
 # -*- coding: utf-8 -*-
 # (c) 2013,2014 Andreas Motl, Elmyra UG
 import logging
+from elmyra.ip.util.config import read_config, read_list, asbool
+from elmyra.ip.util.data.container import SmartBunch
 from pyramid.threadlocal import get_current_request
 from pyramid.url import route_url
 from akhet.urlgenerator import URLGenerator as ApplicationURLGenerator
@@ -14,11 +16,46 @@ log = logging.getLogger(__name__)
 
 def includeme(config):
     """Configure all application-specific subscribers."""
+    config.add_subscriber(global_config, "pyramid.events.ApplicationCreated")
     config.add_subscriber(create_url_generators, "pyramid.events.ContextFound")
     config.add_subscriber(create_tools, "pyramid.events.ContextFound")
     config.add_subscriber(add_renderer_globals, "pyramid.events.BeforeRender")
     config.add_subscriber(add_html_foundation, "pyramid.events.BeforeRender")
 
+def global_config(event):
+    registry = event.app.registry
+    pyramid_settings = registry.settings
+
+    # Read configuration file to get global settings
+    # TODO: Optimize: Only read once, not on each request!
+    application_settings = read_config(pyramid_settings['CONFIG_FILE'])
+
+    # Provide Paste configuration via registry object
+    registry.application_settings = application_settings
+
+
+    # Compute datasource settings
+    datasource_settings = SmartBunch({
+        'datasources': [],
+        'datasource': SmartBunch(),
+        'total': SmartBunch.bunchify({'fulltext_countries': [], 'details_countries': []}),
+        })
+
+    # Read system settings from configuration
+    datasource_settings.datasources = read_list(application_settings.get('ip_navigator', {}).get('datasources'))
+    for datasource in datasource_settings.datasources:
+        application_settings_key = 'datasource_{name}'.format(name=datasource)
+        datasource_info = application_settings.get(application_settings_key, {})
+        datasource_info['fulltext_enabled'] = asbool(datasource_info.get('fulltext_enabled', False))
+        datasource_info['fulltext_countries'] = read_list(datasource_info.get('fulltext_countries', ''))
+        datasource_info['details_enabled'] = asbool(datasource_info.get('details_enabled', False))
+        datasource_info['details_countries'] = read_list(datasource_info.get('details_countries', ''))
+        datasource_settings.datasource[datasource] = SmartBunch.bunchify(datasource_info)
+
+        # Aggregate data for all countries
+        datasource_settings.total.fulltext_countries += datasource_info['fulltext_countries']
+
+    registry.datasource_settings = datasource_settings
 
 def create_url_generators(event):
     """A subscriber for ``pyramid.events.ContextFound`` events. I create various